fix(masterads): corrige las migraciones que rompían la instalación
FKs a tablas inexistentes, índice duplicado y formato de version.yaml.

+++ plugins/aero/masterads/updates/create_workspaces_table.php
@@ -23,7 +23,7 @@ class CreateWorkspacesTable extends \October\Rain\Database\Updates\Migration
             $table->bigIncrements('id');
             $table->string('name', 120);
             $table->string('slug', 120)->unique();
-            $table->unsignedBigInteger('owner_id');
+            $table->unsignedInteger('owner_id');
             $table->json('settings')->nullable();
             $table->timestamps();
